CloudFunctions_recorder/requirements.txt
# Function dependencies, for example:
# package>=version
redis
pyyaml
